A handful of more involved tests.  Surprisingly few things needed fixing.
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/BRANCHES/WCM-DEV2/root@2962 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/avm/AVMService.java
@@ -138,7 +138,7 @@ public interface AVMService
      * @param repName The name of the repository.
      * @return The latest version id of the repository.
      */
-    public long getLatestVersionID(String repName);
+    public int getLatestVersionID(String repName);
     
     /**
      * Snapshot the given repositories.  When this is called everything that has been added,
